two fixtures moved from utils to conftest
